Thank-you page: wider ticket with summary and steps side by side, compact summary on phones

+++ js/script.js
@@ -161,7 +161,7 @@ if (passRef) {
       const m = /^(\d{4})-(\d{2})-(\d{2})$/.exec(req.date || '');
       if (m) {
         when = new Date(+m[1], +m[2] - 1, +m[3]).toLocaleDateString(NL ? 'nl-NL' : 'en-GB',
-          { weekday: 'long', day: 'numeric', month: 'long', year: 'numeric' });
+          { weekday: 'short', day: 'numeric', month: 'long', year: 'numeric' });
       }
       set('sumRef', ref);
       set('sumArtist', req.artist);
